PLANET-5189 ImageArchive: move state to ArchivePicker
